+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next version
 - Fix "Inspect node pools" macro when using managed node groups
+- Add option to install Metrics Server
 
 ## Version 1.0.7 - Bugfix release
 - Add capability to assume IAM role on all cluster operation